Reduced duplicate code with a common query abstract class
