+++ pstatus-notifications-workflow-ktor/src/main/kotlin/query/UploadDigestCountsQuery.kt
@@ -1,15 +1,13 @@
 package gov.cdc.ocio.processingnotifications.query
 
-import gov.cdc.ocio.database.QueryBuilder
 import gov.cdc.ocio.database.persistence.ProcessingStatusRepository
 import gov.cdc.ocio.processingnotifications.workflow.digestcounts.UploadDigestResponse
-import gov.cdc.ocio.types.InstantRange
 import java.time.LocalDate
 
 
 class UploadDigestCountsQuery(
     repository: ProcessingStatusRepository
-): QueryBuilder(repository.reportsCollection) {
+): CommonQuery(repository) {
 
     private fun build(
         utcDateToRun: LocalDate,
@@ -17,14 +15,6 @@ class UploadDigestCountsQuery(
         dataStreamRoutes: List<String>,
         jurisdictions: List<String>
     ): String {
-        val instantRange = InstantRange.fromLocalDate(utcDateToRun)
-        val startTime = timeFunc(instantRange.start.epochSecond)
-        val endTime = timeFunc(instantRange.endInclusive.epochSecond)
-
-        val jurisdictionIdsList = listForQuery(jurisdictions)
-        val dataStreamIdsList = listForQuery(dataStreamIds)
-        val dataStreamRoutesList = listForQuery(dataStreamRoutes)
-
         val querySB = StringBuilder()
 
         querySB.append("""
@@ -38,29 +28,13 @@ class UploadDigestCountsQuery(
             
             FROM $collectionName $cVar
             WHERE ${cPrefix}stageInfo.action IN ${openBkt}'upload-started', 'upload-completed', 'blob-file-copy'${closeBkt}
-                AND ${cPrefix}stageInfo.status = 'SUCCESS'
-                """)
+            """)
 
-        if (dataStreamIds.isNotEmpty())
-            querySB.append("""
-                AND ${cPrefix}dataStreamId IN ${openBkt}$dataStreamIdsList${closeBkt}
-                """)
-
-        if (dataStreamRoutesList.isNotEmpty())
-            querySB.append("""
-                AND ${cPrefix}dataStreamRoute IN ${openBkt}$dataStreamRoutesList${closeBkt}
-                """)
-
-        if (jurisdictionIdsList.isNotEmpty())
-            querySB.append("""
-                AND ${cPrefix}jurisdiction IN ${openBkt}$jurisdictionIdsList${closeBkt}
-                """)
+        querySB.append(whereClause(utcDateToRun, dataStreamIds, dataStreamRoutes, jurisdictions))
 
         querySB.append("""
-                AND ${cPrefix}dexIngestDateTime >= $startTime
-                AND ${cPrefix}dexIngestDateTime < $endTime
-                GROUP BY r.dataStreamId, r.dataStreamRoute, r.jurisdiction
-                """)
+            GROUP BY r.dataStreamId, r.dataStreamRoute, r.jurisdiction
+            """)
 
         return querySB.toString().trimIndent()
     }
