+++ pstatus-notifications-workflow-ktor/src/main/kotlin/query/UploadMetricsQuery.kt
@@ -1,9 +1,7 @@
 package gov.cdc.ocio.processingnotifications.query
 
-import gov.cdc.ocio.database.QueryBuilder
 import gov.cdc.ocio.database.persistence.ProcessingStatusRepository
 import gov.cdc.ocio.processingnotifications.workflow.digestcounts.UploadMetrics
-import gov.cdc.ocio.types.InstantRange
 import java.time.LocalDate
 
 
@@ -14,7 +12,7 @@ import java.time.LocalDate
  */
 class UploadMetricsQuery(
     repository: ProcessingStatusRepository
-): QueryBuilder(repository.reportsCollection) {
+): CommonQuery(repository) {
 
     private fun build(
         utcDateToRun: LocalDate,
@@ -22,14 +20,6 @@ class UploadMetricsQuery(
         dataStreamRoutes: List<String>,
         jurisdictions: List<String>
     ): String {
-        val instantRange = InstantRange.fromLocalDate(utcDateToRun)
-        val startTime = timeFunc(instantRange.start.epochSecond)
-        val endTime = timeFunc(instantRange.endInclusive.epochSecond)
-
-        val jurisdictionIdsList = listForQuery(jurisdictions)
-        val dataStreamIdsList = listForQuery(dataStreamIds)
-        val dataStreamRoutesList = listForQuery(dataStreamRoutes)
-
         val querySB = StringBuilder()
 
         querySB.append("""
@@ -80,28 +70,9 @@ class UploadMetricsQuery(
             
                 FROM $collectionName $cVar
                 WHERE ${cPrefix}stageInfo.action IN ${openBkt}'upload-started', 'upload-completed', 'upload-status'${closeBkt}
-                    AND ${cPrefix}stageInfo.status = 'SUCCESS'
             """)
 
-        if (dataStreamIds.isNotEmpty())
-            querySB.append("""
-                    AND ${cPrefix}dataStreamId IN ${openBkt}$dataStreamIdsList${closeBkt}
-            """)
-
-        if (dataStreamRoutesList.isNotEmpty())
-            querySB.append("""
-                    AND ${cPrefix}dataStreamRoute IN ${openBkt}$dataStreamRoutesList${closeBkt}
-            """)
-
-        if (jurisdictionIdsList.isNotEmpty())
-            querySB.append("""
-                    AND ${cPrefix}jurisdiction IN ${openBkt}$jurisdictionIdsList${closeBkt}
-            """)
-
-        querySB.append("""
-                    AND ${cPrefix}dexIngestDateTime >= $startTime
-                    AND ${cPrefix}dexIngestDateTime < $endTime
-        """)
+        querySB.append(whereClause(utcDateToRun, dataStreamIds, dataStreamRoutes, jurisdictions))
 
         querySB.append("""
                 GROUP BY ${cPrefix}uploadId
